add pertemuan11

+++ kuliah/pertemuan11/detail.php
@@ -23,7 +23,7 @@ $m = query("SELECT * FROM mahasiswa WHERE id = $id");
     <li>Nama : <?= $m['nama']; ?></li>
     <li>Email : <?= $m['email']; ?></li>
     <li>Jurusan : <?= $m['jurusan']; ?></li>
-    <li><a href="">ubah</a> | <a href="hapus.php?id=<?= $m['id']; ?>">hapus</a></li>
+    <li><a href="ubah.php?id=<?= $m['id']; ?>">ubah</a> | <a href="hapus.php?id=<?= $m['id']; ?>" onclick="return confirm('Apakah Anda yakin?');">hapus</a></li>
     <li><a href="index.php">Kembali ke daftar mahasiswa</a></li>
   </ul>
 </body>
